moved xacro utils to the rsl/env
